eccezioni aggiunte ai Dao: utente, acquisto e preferenza

+++ src/main/java/data/DAOAcquisto/AcquistoDAO.java
@@ -1,5 +1,9 @@
 package data.DAOAcquisto;
 
+import data.Exceptions.OggettoGiaPresenteException;
+import data.Exceptions.OggettoNonCancellatoException;
+import data.Exceptions.OggettoNonInseritoException;
+import data.Exceptions.OggettoNonTrovatoException;
 import data.utils.Dao;
 import data.utils.SingletonJDBC;
 import data.DAOCanzone.Canzone;
@@ -19,11 +23,15 @@ public class AcquistoDAO implements AcquistoAPI {
      * @return un oggetto Acquisto
      * */
     public Acquisto doGet(String chiave) throws SQLException {
+        if(chiave == null || !chiave.contains(";"))
+            throw new IllegalArgumentException("la chiave è null o non valida");
         String chiavi[] = chiave.split(";");
         PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement("SELECT * FROM acquisto WHERE username=? && codCanzone=?");
         preparedStatement.setString(1,chiavi[0]);
         preparedStatement.setString(2,chiavi[1]);
         ResultSet resultSet = preparedStatement.executeQuery();
+        if(resultSet.getRow()==0)
+            throw new OggettoNonTrovatoException("L'acquisto non è stato trovato: "+chiave);
         return new Acquisto(resultSet.getString("username"),resultSet.getString("codCanzone"));
     }
 
@@ -32,40 +40,67 @@ public class AcquistoDAO implements AcquistoAPI {
      * @param acquisto l'oggetto da salvare. Prima di invocare il metodo bisogna settare la username e il codCanzone all'interno dell oggetto
      * @return true se il salvataggio avviene correttamente, false altrimenti
      * */
-    public boolean doSave(Acquisto acquisto) throws SQLException {
-        if(acquisto.getCodCanzone() == null || acquisto.getUsename() == null)
+    public void doSave(Acquisto acquisto) throws SQLException {
+        if(acquisto == null || acquisto.getCodCanzone() == null || acquisto.getUsename() == null)
             throw new IllegalArgumentException("parametri non settati all'interno di acquisto");
-
-       PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement("INSERT INTO uc VALUES (?,?)");
-       preparedStatement.setString(1,acquisto.getUsename());
-       preparedStatement.setString(2,acquisto.getCodCanzone());
-       return preparedStatement.executeUpdate()==1;
+        try{
+            doGet(acquisto.getCodCanzone()+";"+acquisto.getUsename());
+            throw new OggettoGiaPresenteException("L'acquisto è gia presente");
+        }catch (OggettoNonTrovatoException e) {
+            PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement("INSERT INTO uc VALUES (?,?)");
+            preparedStatement.setString(1, acquisto.getUsename());
+            preparedStatement.setString(2, acquisto.getCodCanzone());
+            if (preparedStatement.executeUpdate() != 1)
+                throw new OggettoNonInseritoException("L'acquisto non è stato inserito");
+        }
     }
 
     @Override
     /**Questo metodo cancella un acquisto dal DB
      * @param chiave la concatenazione della username e del codice canzone. Esempio: "pluto;C94"e
-     * @return true se l'acquisto è stato cancellato, false altrimenti
+     * @return void
      * */
-    public boolean doDelete(String chiave) throws SQLException {
-        String[] chiavi = chiave.split(";");
-        PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement("DELETE FROM uc WHERE username=? && codCanzone=?");
-        preparedStatement.setString(1,chiavi[0]);
-        preparedStatement.setString(2, chiavi[1]);
-        return preparedStatement.executeUpdate()==1;
+    public void doDelete(String chiave) throws SQLException {
+        if(chiave == null || !chiave.contains(";"))
+            throw new IllegalArgumentException("la chiave è null o non valida");
+        try{
+            doGet(chiave);
+            String[] chiavi = chiave.split(";");
+            PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement("DELETE FROM uc WHERE username=? && codCanzone=?");
+            preparedStatement.setString(1,chiavi[0]);
+            preparedStatement.setString(2, chiavi[1]);
+        if(preparedStatement.executeUpdate()!=1)
+            throw new OggettoNonCancellatoException("l'acquisto non è stato cancellato");
+        }catch (OggettoNonTrovatoException e){
+            throw e;
+        }
     }
 
-    /**Inserisce una canzone acquistata*/
-    public boolean doInsertCanzoneAcquistata(String username, String codice) throws SQLException {
-        PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement(CanzoneQuery.getQueryDoInsertCanzoneAcquistata());
-        preparedStatement.setString(1,username);
-        preparedStatement.setString(2,codice);
-        return preparedStatement.executeUpdate()==1;
+    @Override
+    /**Questo metodo cancella un acquisto dal DB
+     * @param chiave la concatenazione della username e del codice canzone. Esempio: "pluto;C94"e
+     * @return void
+     * */
+    public void doInsertCanzoneAcquistata(String username, String codice) throws SQLException {
+        if(username == null || codice == null)
+            throw new IllegalArgumentException("username o codice sono null");
+        try {
+            doGet(codice + ";" + username);
+            throw new OggettoGiaPresenteException("L'acquisto è gia presente");
+        }catch(OggettoNonTrovatoException e){
+                PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement(CanzoneQuery.getQueryDoInsertCanzoneAcquistata());
+                preparedStatement.setString(1, username);
+                preparedStatement.setString(2, codice);
+                if (preparedStatement.executeUpdate() != 1)
+                    throw new OggettoNonInseritoException("La canzone non è stata acquistata");
+            }
     }
 
     /**Ritorna la lista delle canzoni acquistate dall utente*/
     public List<String> doRetrieveCodiciCanzoniAcquistate(String username) throws SQLException {
-
+        if(username == null){
+            throw new IllegalArgumentException("username è null");
+        }
         PreparedStatement preparedStatement = SingletonJDBC.getConnection().prepareStatement(CanzoneQuery.getQueryDoRetrieveCodiciCanzoniAcquistate());
         preparedStatement.setString(1,username);
         preparedStatement.setString(2,username);
@@ -78,6 +113,9 @@ public class AcquistoDAO implements AcquistoAPI {
 
     /**Ritorna le canzoni acquistate dall'utente*/
     public List<Canzone> doRetrieveCanzoniAcquistate(String username) throws SQLException {
+        if(username == null){
+            throw new IllegalArgumentException("username è null");
+        }
         PreparedStatement statement = SingletonJDBC.getConnection().prepareStatement(CanzoneQuery.getQueryDoRetrieveCanzoniAcquistate());
         statement.setString(1,username);
         statement.setString(2, username);
